some-changes

+++ app/Http/Controllers/UserController.php
@@ -5,66 +5,60 @@ namespace App\Http\Controllers;
 use App\Http\ApiResponse;
 use App\Models\User;
 use Illuminate\Http\Request;
-Use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
 {
-    public function registerUser(Request $req){
+    public function registerUser(Request $req)
+    {
 
         $req->validate([
-            "name"=> "required",
-            "email"=> "email|required|unique:users",
-            "password"=> "required"
+            "name" => "required",
+            "email" => "email|required|unique:users",
+            "password" => "required"
         ]);
-        
+
         $name = $req->name;
         $email = $req->email;
         $password = $req->password;
-        
+
         $user =  new User();
-    
+
         $user->name = $name;
         $user->email = $email;
         $user->password = Hash::make($password);
         $user->save();
-        
-        $user->assignRole('employee');
-        
 
-        $data = [
-            'message' => 'User Registerd Successfully',
-            'code' => 200,
-            'data' => $user,  
-        ];
+        unset($user['roles']);
 
-        return response()->json($data);
+        return ApiResponse::success($user);
     }
 
-    public function login(Request $req){
+    public function login(Request $req)
+    {
 
         $email = $req->email;
         $password = $req->password;
         $user = User::where(['email' => $email])->first();
 
-        if(!$user)
-        {
+        if (!$user) {
             return response([
                 "code" => "404",
-                "message"=> "User not found"
+                "message" => "User not found"
             ]);
         }
 
-        if(!Hash::check($password, $user->password))
-        {
+        if (!Hash::check($password, $user->password)) {
             return response([
-                "message"=>"Incorrect Password"
+                "message" => "Incorrect Password"
             ]);
         }
         $roles = $user->getRoleNames();
         foreach ($user->tokens as $tKey => $token) {
             $token->delete();
         }
-        
+
         unset($user['tokens']);
         unset($user['roles']);
 
@@ -72,74 +66,30 @@ class UserController extends Controller
         $token = $user->createToken($user->name)->plainTextToken;
 
         $data = [
-            'message' => 'User logged in Successfully',
-            'code' => 200,
-            'data' => $user,
+            'user' => $user,
             'token' => $token
         ];
-
-        return response()->json($data);
+        return ApiResponse::success($data);
     }
 
-    public function logout(Request $request){
-
-        $request->user()->tokens()->delete();
-
-        return response([
-            "code" => 200,
-            "message" => "User Logged out successfully"
-        ]);
-    }
-
-    public function userDetails(Request $req){
-        
-        $email = $req->email;
-        $user = User::where(['email' => $email])->with(['topics'])->first();
-
-        if(!$user)
-        {
-            return response([
-                "code" => "404",
-                "message"=> "User not found"
-            ]);
+    public function logout(Request $request)
+    {
+        if ($request->user()) {
+            $request->user()->tokens()->delete();
         }
 
-        $data = [
-            'message' => 'Success',
-            'code' => 200,
-            'data' => $user,
-        ];
-
-        return response()->json($data);
+        return ApiResponse::success("User Logged out successfully");
     }
 
-    public function editDetials(Request $req){
+    public function userDetails(Request $req)
+    {
 
-        $req->validate([
-            "id" => "required"
-        ]);
-        
-        $id = $req->id;
-        $name = $req->name;
-        $email = $req->email;
 
-        $user = User::find($id);
+        $user = User::all()->except(Auth::id());
 
-        if (!$user){
-            return response([
-                "code" => 404,
-                "message" => "user not found"
-            ]);
+        if (!$user) {
+            return ApiResponse::errorGeneral("user not found");
         }
-
-        $user->name = $name;
-        $user->email = $email;
-        $user->save();
-        
-        return response([
-            "code" => 200,
-           "message" => "User Details Edited Successfully",
-           "user" => $user 
-        ]);
+        return ApiResponse::success($user);
     }
 }
